pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name="eneres_100_hw_libraries"
dynamic = ["version"]
version="0.1.3"
description="Python library with tools for UC Berkeley's ENE,RES 100"
authors = [
    {name = "Serrindipity", email = "jon_quitoriano@berkeley.edu"},
]
readme='README.md'
dependencies=[
    "eneres_100_unum @ git+https://github.com/Serrindipity/ENERES-100-Unum.git@main"
]

[project.urls]
Repository = "https://github.com/Serrindipity/ENERES-100-HW-Libraries"

[tool.setuptools.packages]
find = {}  # Scan the project directory with the default parameters